Customize org-mode and org-projectile
Also set the base action for buffers to reuse, then dispaly in the same
window. Help buffers and links in help buffers tend to open in other
windows. Aside from display-buffer-base-action, there is also
same-window-buffer-names. The current configuration of this variable
isn't necessary given display-buffer-base-action, but I am keeping it as
an example of the flexibility of emacs.

Also uncomment ranger package.

Modify org-open-at-point to do what I want, similar to my which-key
hacks.